Handle comments and whitespace around braces
This commit addresses blacktex issue:
https://github.com/nschloe/blacktex/issues/13

Comments
--------

The percent signs at EOL are not treated as a comments.
It is critical for TikZ code, for example.
This snippet will thus not be changed.
```
\DeclareOption{alignedleftspaceyesifneg}{%
\newcommand{\alignedspace}@left{%
  \edef\@tempa{\expandafter\@car\the\lastskip\@nil}%
  \if-\@tempa\null\,%
  \else
    \edef\@tempa{\expandafter\@car\the\lastkern\@nil}%
    \if-\@tempa\null\,%
    \else\null
    \fi
  \fi}%
}
```

Percents preceded by backslash are not treated as comments as well.

Brackets/parens/curly braces
----------------------------

Whitespace clean-up around brackets/parens/curly_braces is changed.
If the new line is encountered inside, then do not change anything.
It may be beneficial for long image filenames and/or captions.
This snippet will thus not be changed.

```
\includegraphics{
    long_path_to_image_file_on_new_line
}
```